B Tool Demos/CloudPubSubFun/requirements.txt
pandas
tweepy
google-cloud-pubsub